Set settings.
