+++ .travis.yml
@@ -5,10 +5,16 @@ go:
     - tip
 
 before_install:
-    - go get -t -v ./...
+    - sudo add-apt-repository ppa:masterminds/glide -y
+    - sudo apt-get update -q
+    - sudo apt-get install glide -y
+    - mkdir -p $GOPATH/bin
+
+install:
+    - make deps
 
 script:
-    - ./go.test.sh
+    - make ci-test
 
 after_success:
     - bash <(curl -s https://codecov.io/bash)
