+++ cli_test.go
