--- main.go
@@ -1,13 +1,0 @@
-// Copyright © 2016 budougumi0617 All rights reserved.
-
-// The Tasks creates Task List Comments from files.
-package main
-
-import "os"
-
-const Version string = "v0.0.1"
-
-func main() {
-	cli := &CLI{outStream: os.Stdout, errStream: os.Stderr}
-	os.Exit(cli.Run(os.Args))
-}
